+++ clock.h
@@ -4,5 +4,5 @@ void initialize_clock(void);
 void delay1s(void);
 void delay_ms(uint32_t t);
 
-uint32_t get_time_ms(void);
+uint16_t get_time_ms(void);
